refactor(executor): split cost/auth out of codex.go and claude.go
The provenance and token-gate work pushed both executors past the 800-line
check-file-sizes gate (codex 817, claude 843) — a CI failure, not a warning.

Extracted into a cost.go per package: authLane() + CostModel() for both, plus
getModel() for claude. These belong together — what a run costs and whether
anyone was charged for it are the same question asked twice.

codex 817 -> 761, claude 843 -> 778. No behaviour change.
