+++ internal/executor/claude/claude.go
@@ -754,38 +754,6 @@ func (e *ClaudeExecutor) Capabilities() []executor.Capability {
 	}
 }
 
-// authLane reports whether claude runs are charged per token.
-//
-// Mirrors the M-CLOUD-DUAL-AUTH branch in Execute: AILANG_AUTH_MODE=apikey means
-// ANTHROPIC_API_KEY drives a metered account; anything else is the OAuth
-// subscription lane, where the CLI still emits a non-zero total_cost_usd that
-// nobody is charged. On the eval rig the key is deliberately stripped, so the
-// default is the common case, not an edge case.
-func (e *ClaudeExecutor) authLane() executor.AuthLane {
-	if os.Getenv("AILANG_AUTH_MODE") == "apikey" {
-		return executor.AuthLaneBilled
-	}
-	return executor.AuthLaneSubscription
-}
-
-// CostModel returns pricing information for cost calculations.
-//
-// NOT used for Result.CostUSD: the claude CLI reports its own
-// total_cost_usd and the executor banks that figure directly. Kept because
-// the Executor interface requires it and callers may use it for pre-flight
-// estimates. Audited 2026-07-30 — do not assume this table is what gets
-// banked. Note the CLI's figure is itself a list-price equivalent when the
-// rig authenticates via OAuth subscription, not metered spend.
-func (e *ClaudeExecutor) CostModel() *executor.CostModel {
-	// Default to Haiku pricing
-	return &executor.CostModel{
-		ProviderName:    "anthropic",
-		InputTokenCost:  0.001,  // $1.00 per 1M
-		OutputTokenCost: 0.005,  // $5.00 per 1M
-		CacheReadCost:   0.0001, // $0.10 per 1M
-	}
-}
-
 // HealthCheck verifies the executor is configured and accessible
 func (e *ClaudeExecutor) HealthCheck(ctx context.Context) error {
 	// Check if claude binary exists
@@ -799,39 +767,6 @@ func (e *ClaudeExecutor) HealthCheck(ctx context.Context) error {
 	return nil
 }
 
-// Close releases any resources held by the executor
-func (e *ClaudeExecutor) Close() error {
-	return nil
-}
-
-func (e *ClaudeExecutor) getModel(task *executor.Task) string {
-	if task.Model != "" {
-		return task.Model
-	}
-	return e.model
-}
-
-// claudeHeadlessResult matches Claude CLI output structure
-type claudeHeadlessResult struct {
-	Type         string      `json:"type"`
-	Subtype      string      `json:"subtype"`
-	IsError      bool        `json:"is_error"`
-	Result       string      `json:"result"`
-	NumTurns     int         `json:"num_turns"`
-	DurationMS   int         `json:"duration_ms"`
-	TotalCostUSD float64     `json:"total_cost_usd"`
-	SessionID    string      `json:"session_id"`
-	Usage        claudeUsage `json:"usage"`
-}
-
-type claudeUsage struct {
-	InputTokens              int `json:"input_tokens"`
-	OutputTokens             int `json:"output_tokens"`
-	CacheReadInputTokens     int `json:"cache_read_input_tokens"`
-	CacheCreationInputTokens int `json:"cache_creation_input_tokens"`
-}
-
-// Register registers the Claude executor with the global factory
 func Register() {
 	executor.GlobalFactory().Register("claude", func(cfg *executor.Config) (executor.Executor, error) {
 		return New(cfg)
